add no-elif-return check
