+++ gdtoolkit/linter/misc_checks.py
@@ -1,19 +1,22 @@
-from functools import partial
 from types import MappingProxyType
 from typing import List
 
 from lark import Tree
 
 from .problem import Problem
-from .no_else_return_check import no_else_return_check
+from .if_return_checks import no_elif_return_check, no_else_return_check
 
 
 def lint(parse_tree: Tree, config: MappingProxyType) -> List[Problem]:
     disable = config["disable"]
     checks_to_run_w_tree = [
         (
+            "no-elif-return",
+            no_elif_return_check,
+        ),
+        (
             "no-else-return",
-            partial(no_else_return_check, config["no-else-return"]),
+            no_else_return_check,
         ),
     ]
     problem_clusters = map(
